backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pdfplumber>=0.10.0
pydantic>=2.5.0
python-multipart>=0.0.6
pdf2image
Pillow
google-genai
python-dotenv

